Save! Instead of Save
Try to call `save!` on objects instead of `save`. A Sham should be
configured with valid attributes, if it ever fails it should fail loudly
instead of returning an invalid object. Note: this does not impact
`:build` calls - those will not save unless explicitly asked to do so by
the calling code.

Fixes #6